// src/api/generated/models/activityItemEventType.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ActivityItemEventType = typeof ActivityItemEventType[keyof typeof ActivityItemEventType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActivityItemEventType = {
  QUOTE_LINE_SELECTED: 'QUOTE_LINE_SELECTED',
  QUOTE_LINE_SELECTION_UPDATED: 'QUOTE_LINE_SELECTION_UPDATED',
  QUOTE_LINE_UNSELECTED: 'QUOTE_LINE_UNSELECTED',
  INVOICE_CREATED: 'INVOICE_CREATED',
  INVOICE_LINE_ADDED: 'INVOICE_LINE_ADDED',
  INVOICE_LINE_UPDATED: 'INVOICE_LINE_UPDATED',
  INVOICE_LINE_REMOVED: 'INVOICE_LINE_REMOVED',
  INVOICE_ISSUED: 'INVOICE_ISSUED',
  INVOICE_CONFIRMED: 'INVOICE_CONFIRMED',
  PAYMENT_REGISTERED: 'PAYMENT_REGISTERED',
  PAYMENT_UPLOADED: 'PAYMENT_UPLOADED',
  PAYMENT_CONFIRMED: 'PAYMENT_CONFIRMED',
  PAYMENT_REJECTED: 'PAYMENT_REJECTED',
  SHIPPING_INVOICE_CREATED: 'SHIPPING_INVOICE_CREATED',
  SHIPPING_INVOICE_ISSUED: 'SHIPPING_INVOICE_ISSUED',
  SHIPPING_INVOICE_IN_TRANSIT: 'SHIPPING_INVOICE_IN_TRANSIT',
  SHIPPING_INVOICE_DELIVERED: 'SHIPPING_INVOICE_DELIVERED',
  SHIPPING_LINE_ADDED: 'SHIPPING_LINE_ADDED',
  SHIPPING_LINE_UPDATED: 'SHIPPING_LINE_UPDATED',
  SHIPPING_LINE_REMOVED: 'SHIPPING_LINE_REMOVED',
  CONSOLIDATION_CREATED: 'CONSOLIDATION_CREATED',
  CONSOLIDATION_PLANNED: 'CONSOLIDATION_PLANNED',
  CONSOLIDATION_IN_TRANSIT: 'CONSOLIDATION_IN_TRANSIT',
  CONSOLIDATION_CUSTOMS: 'CONSOLIDATION_CUSTOMS',
  CONSOLIDATION_DELIVERED: 'CONSOLIDATION_DELIVERED',
  CONSOLIDATION_SHIPPING_INVOICE_ADDED: 'CONSOLIDATION_SHIPPING_INVOICE_ADDED',
  CONSOLIDATION_SHIPPING_INVOICE_REMOVED: 'CONSOLIDATION_SHIPPING_INVOICE_REMOVED',
  REQUEST_IMPORTED_FROM_INTEGRATION: 'REQUEST_IMPORTED_FROM_INTEGRATION',
  REQUEST_UPDATED_FROM_INTEGRATION: 'REQUEST_UPDATED_FROM_INTEGRATION',
} as const;